Elli: clarify which wallboxes are supported best (#9036)
The current Elli wallboxes contain lots of software bugs, which make their usage and also support a nightmare.

Only the following two scenarios provide a somehow stable environment which can be “recommended”:

- Elli Pro: No CT coil usage, LAN connection. If a LAN connection is not possible, addition of an external meter
- Elli Pro: No CT coil usage, external meter usage.

A LAN connection is always highly recommended, as even the slightest network connection drop will allow charging of the car with minimal power until reconnection is establied.

There is no workaround for the wallbox software crashing often on connection attempts which can take a few minutes for the connection to work. During that time the EV is allowed to charge with minimal power.