Planner: fix estimator disabled when soc provided by charger (#10698)
When the vehicles SoC is provided by the charger, the planner currently assumes that the EV is not charged at all. This can be reproduced by setting a target Soc to be identical to the current, and the planner will show a charge time that would be needed to reach the current Soc.

The reason is `s.vehicleSoc` in `estimator.go` not being set as `publichSocAndRange()` exits before setting it.